# api/handler/download.go
package handler

import (
	"fmt"
	"net/http"
	"path/filepath"
	"sort"

	"github.com/gin-gonic/gin"
	"github.com/qjfoidnh/BaiduPCS-Go/api/model"
	"github.com/qjfoidnh/BaiduPCS-Go/baidupcs"
	"github.com/qjfoidnh/BaiduPCS-Go/baidupcs/pcserror"
	"github.com/qjfoidnh/BaiduPCS-Go/internal/pcscommand"
	"github.com/qjfoidnh/BaiduPCS-Go/internal/pcsconfig"
	"github.com/qjfoidnh/BaiduPCS-Go/internal/pcsfunctions/pcsdownload"
	"github.com/qjfoidnh/BaiduPCS-Go/pcsutil/converter"
	"github.com/qjfoidnh/BaiduPCS-Go/pcsutil/taskframework"
	"github.com/qjfoidnh/BaiduPCS-Go/requester/downloader"
	"github.com/qjfoidnh/BaiduPCS-Go/requester/transfer"
)

// Locate 获取下载直链
// Locate 获取下载直链
// @Summary 获取下载直链
// @Description 获取指定文件的下载直链
// @Tags 上传下载
// @Accept json
// @Produce json
// @Param request body model.DownloadRequest true "请求内容"
// @Success 200 {object} model.Response
// @Failure 400 {object} model.Response
// @Failure 500 {object} model.Response
// @Router /api/locate [post]
func Locate(c *gin.Context) {
	var req model.DownloadRequest // 复用 DownloadRequest，只用 Paths
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, model.ErrorResponse(400, err.Error()))
		return
	}

	paths, err := matchPaths(req.Paths...)
	if err != nil {
		c.JSON(http.StatusInternalServerError, model.ErrorResponse(500, err.Error()))
		return
	}

	pcs := pcscommand.GetBaiduPCS()
	var links []map[string]interface{}

	// 获取文件ID (fid)
	for _, p := range paths {
		f, err := pcs.FilesDirectoriesMeta(p)
		if err != nil {
			links = append(links, map[string]interface{}{
				"path":  p,
				"error": err.Error(),
			})
			continue
		}

		// 获取下载链接
		// Func: LocateDownload(pcspath string) (info *URLInfo, pcsError pcserror.Error)
		info, err := pcs.LocateDownload(p)
		if err != nil {
			links = append(links, map[string]interface{}{
				"path":  p,
				"error": err.Error(),
			})
			continue
		}

		// 选取最佳链接（通常是第一个）
		urlStr := ""
		u := info.SingleURL(true) // true for HTTPS
		if u != nil {
			urlStr = u.String()
		}

		links = append(links, map[string]interface{}{
			"path":     p,
			"fs_id":    f.FsID,
			"url":      urlStr,
			"urls":     info.URLs, // 包含所有备选链接
			"filename": f.Filename,
			"size":     f.Size,
		})
	}

	c.JSON(http.StatusOK, model.SuccessResponse(gin.H{
		"links": links,
	}))
}

// Download 下载文件到服务器本地
// Download 下载文件到服务器本地
// @Summary 下载文件
// @Description 将网盘文件下载到服务器本地
// @Tags 上传下载
// @Accept json
// @Produce json
// @Param request body model.DownloadRequest true "下载请求"
// @Success 200 {object} model.Response
// @Failure 400 {object} model.Response
// @Router /api/download [post]
func Download(c *gin.Context) {
	var req model.DownloadRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, model.ErrorResponse(400, err.Error()))
		return
	}

	// 1. 匹配路径
	paths, err := matchPaths(req.Paths...)
	if err != nil {
		c.JSON(http.StatusInternalServerError, model.ErrorResponse(500, err.Error()))
		return
	}

	// 2. 配置选项
	// 为了简化 API，使用默认配置或部分可配置
	parallel := pcsconfig.Config.MaxParallel
	saveTo := req.SaveTo

	// 如果没有指定保存路径，使用默认下载路径
	if saveTo == "" {
		saveTo = pcsconfig.Config.ActiveUser().Workdir // 或者 activeUser.SavePath?
		// CLI 默认逻辑是 GetSavePath(p)
		// 这里我们暂时设为空，让后续逻辑处理
	} else {
		// 如果指定了绝对路径，确保存在
		// 简单处理，不做过多检查
	}

	pcs := pcscommand.GetBaiduPCS()

	// 3. 收集文件信息 (复用 RunDownload 逻辑的一部分)
	var fileDirList []*baidupcs.FileDirectory

	for _, p := range paths {
		pcs.FilesDirectoriesRecurseList(p, baidupcs.DefaultOrderOptions, func(depth int, _ string, fd *baidupcs.FileDirectory, pcsError pcserror.Error) bool {
			if pcsError != nil {
				return true
			}
			fileDirList = append(fileDirList, fd)
			return true
		})
	}

	// 4. 配置下载器
	cfg := &downloader.Config{
		Mode:      transfer.RangeGenMode_BlockSize,
		CacheSize: pcsconfig.Config.CacheSize,
		BlockSize: baidupcs.InitRangeSize,
		MaxRate:   pcsconfig.Config.MaxDownloadRate,
		TryHTTP:   !pcsconfig.Config.EnableHTTPS,
	}

	// 5. 任务执行器
	executor := taskframework.TaskExecutor{
		IsFailedDeque: true,
	}

	statistic := &pcsdownload.DownloadStatistic{}

	// 排序小文件优先
	sort.Slice(fileDirList, func(i, j int) bool {
		return fileDirList[i].Size < fileDirList[j].Size
	})

	var startedTasks []string

	for _, v := range fileDirList {
		if v.Isdir {
			continue
		}

		// 计算本地保存路径
		var localSavePath string
		if saveTo != "" {
			// 如果指定了 SaveTo，则所有文件下载到该目录下，保持目录结构??
			// 简单起见，如果指定了 SaveTo，则视为根目录
			// 实际上 RunDownload 的逻辑很复杂，涉及 FullPath 选项
			// 这里简化：下载到 SaveTo / Filename
			localSavePath = filepath.Join(saveTo, v.Filename)
		} else {
			// 使用默认保存路径逻辑
			localSavePath = pcsconfig.Config.ActiveUser().GetSavePath(v.Path)
		}

		newCfg := *cfg
		unit := pcsdownload.DownloadTaskUnit{
			Cfg: &newCfg,
			PCS: pcs,
			// VerbosePrinter:     nil, // 设为nil以尝试静默? 代码中可能会 panic 如果调用 Warnf
			// 只能希望它不要打印太多
			PrintFormat:          "", // 禁用打印格式?
			ParentTaskExecutor:   &executor,
			DownloadStatistic:    statistic,
			IsPrintStatus:        false,
			IsExecutedPermission: false,
			IsOverwrite:          req.Save, // 复用 Save 字段表示 overwrite? 不，DownloadRequest.Save 只是 flag
			// 我们假设 overwrite = false
			NoCheck:      pcsconfig.Config.NoCheck,
			DownloadMode: pcsdownload.DownloadModePCS,
			PcsPath:      v.Path,
			FileInfo:     v,
			SavePath:     localSavePath,
		}

		executor.SetParallel(parallel)
		executor.Append(&unit, 3) // MaxRetry 3
		startedTasks = append(startedTasks, fmt.Sprintf("%s -> %s", v.Path, localSavePath))
	}

	if len(startedTasks) == 0 {
		c.JSON(http.StatusBadRequest, model.ErrorResponse(400, "没有可下载的文件"))
		return
	}

	// 6. 执行 (同步执行，可能会阻塞很久!)
	// 建议放入 goroutine?
	// 但如果放入 goroutine，API 立即返回，用户不知道什么时候完成。
	// 为了简单，本次实现为**同步**。如果文件大，会超时。
	// 但这是 "Server-Side Download"，通常希望它是异步的。
	// 可是没有任务管理系统，我无法返回 Task ID 供查询。
	// 妥协：同步执行，但仅建议下载小文件或用于测试。
	// 或者：返回 "Started" 并后台运行。

	// 决定：后台运行，返回 Started。因为这可能需要很久。
	go func() {
		statistic.StartTimer()
		executor.Execute()
		// TODO: 记录日志或回调?
		fmt.Printf("API Download finished: %d files\n", len(startedTasks))
	}()

	c.JSON(http.StatusOK, model.SuccessResponse(gin.H{
		"message":        "下载任务已在后台启动",
		"files":          startedTasks,
		"total_size_str": converter.ConvertFileSize(statistic.TotalSize()), // 此时可能还是0
	}))
}