# docs/swagger.yaml
definitions:
...
  model.DownloadRequest:
    properties:
      paths:
        description: 要下载的路径列表
        items:
          type: string
        minItems: 1
        type: array
      save:
        description: 是否保存到本地
        type: boolean
      save_to:
        description: 保存路径
        type: string
    required:
    - paths
    type: object
  model.ListRequest:
    properties:
      desc:
        description: 是否降序
        type: boolean
      order:
        description: 排序字段：name/time/size
        type: string
      path:
        description: 路径
        type: string
    type: object
...
  model.MetaRequest:
    properties:
      paths:
        description: 路径列表
        items:
          type: string
        minItems: 1
        type: array
    required:
    - paths
    type: object
...
  model.Response:
    properties:
      code:
        description: 状态码：0-成功，非0-失败
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/cd:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 改变当前用户的工作目录
      parameters:
      - description: 目标路径
        in: formData
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 切换工作目录
      tags:
      - 工作目录
...
  /api/download:
    post:
      consumes:
      - application/json
      description: 将网盘文件下载到服务器本地
      parameters:
      - description: 下载请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: 下载文件
      tags:
      - 上传下载
...
  /api/locate:
    post:
      consumes:
      - application/json
      description: 获取指定文件的下载直链
      parameters:
      - description: 请求内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DownloadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取下载直链
      tags:
      - 上传下载
  /api/ls:
    post:
      consumes:
      - application/json
      description: 列出指定路径下的文件和目录
      parameters:
      - description: 列表请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      summary: 列出文件和目录
      tags:
      - 文件管理
  /api/meta:
    post:
      consumes:
      - application/json
      description: 获取指定文件或目录的详细信息
      parameters:
      - description: 元数据请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.MetaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取文件元数据
      tags:
      - 文件管理
...
  /api/pwd:
    get:
      consumes:
      - application/json
      description: 获取当前用户的工作目录路径
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取当前工作目录
      tags:
      - 工作目录
...
  /api/search:
    get:
      consumes:
      - application/json
      description: 在指定目录下搜索文件
      parameters:
      - description: 搜索关键字
        in: query
        name: keyword
        required: true
        type: string
      - description: 搜索路径 (默认根目录)
        in: query
        name: path
        type: string
      - description: 是否递归
        in: query
        name: recurse
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 搜索文件
      tags:
      - 文件管理
...
swagger: "2.0"